+++ scripts/lib/validation/config_updates.py
@@ -8,3 +8,5 @@ Config validation update record for PyPSA-EUR.
 Any imports included in this file will be automatically registered as config updaters and applied to the base config schema.
 See the docs for more details: https://pypsa-eur.readthedocs.io/en/latest/validation_dev.html#soft-fork-ext
 """
+
+import scripts.gb_model._update_config