+++ doc/gb-model/release_notes.rst
@@ -14,6 +14,10 @@ Unreleased
 
 * Add missing carrier for non-networked electrolysis demand (#242).
 * Capitalise all EV carriers (#243).
+* Automate ETYS boundary line/link crossings.
+* Update config filenames. `config/config.gb.2024.yaml` defines the FES2024 configuration; `config/config.default.gb.yaml` is automatically generated and should not be edited directly (#235).
+* Add config validation and associated documentation for the additional configuration required for the gb-dispatch-model (#235).
+* Switch regions `GB 30` and `GB 31` and rename new `GB 31` to `GB NI` to more explicitly represent the Northern Ireland region.
 * Set line capacities to large number for constrained network to make sure only boundary capabilities are limiting (#241).
 * New config option to enable aggregating the PyPSA network time dimensions, to reduce solve times (#229).
 * Created new OSM pre-built network, available at https://zenodo.org/records/18712831, to include updates made in upstream PyPSA-Eur (incl. ignoring unbuilt lines) (#237).